Documentation for serializers
