+++ db-scheduler/src/main/java/com/github/kagkarlsson/scheduler/serializer/SerializerWithFallbackDeserializers.java
@@ -1,0 +1,45 @@
+/**
+ * Copyright (C) Gustav Karlsson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.github.kagkarlsson.scheduler.serializer;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+public class SerializerWithFallbackDeserializers implements Serializer {
+
+    private static final Logger LOG = LoggerFactory.getLogger(SerializerWithFallbackDeserializers.class);
+    private final Serializer serializer;
+    private final Serializer[] fallbackDeserializers;
+
+    public SerializerWithFallbackDeserializers(Serializer serializer, Serializer ... fallbackDeserializers) {
+        this.serializer = serializer;
+        this.fallbackDeserializers = fallbackDeserializers;
+    }
+    @Override
+    public byte[] serialize(Object data) {
+        return serializer.serialize(data);
+    }
+
+    @Override
+    public <T> T deserialize(Class<T> clazz, byte[] serializedData) {
+        try {
+            return serializer.deserialize(clazz, serializedData);
+        } catch (Exception e) {
+            LOG.debug("Failed to deserialize data. Trying fallback Serializer.", e);
+            return fallbackDeserializers[0].deserialize(clazz, serializedData);
+        }
+    }
+}
